+++ packages/radio-button/src/presenters/stylesheet.js
@@ -79,13 +79,17 @@ export default function stylesheet(props, themeData, theme) {
   const opacity = disabled ? themeData["component.disabled.opacity"] : "1.0";
   // because we don't have access to the checked status in react, we need to
   // write css rules to handle both cases
+  const borderColor =
+    hasFocus && !hasHover
+      ? themeData["checkbox.focus.borderColor"]
+      : themeData["checkbox.borderColor"];
   const checkedInnerShadow = `inset 0 0 0 ${
     themeData["checkbox.pressed.halo.width"]
   } ${themeData["checkbox.checked.backgroundColor"]}`;
   const uncheckedInnerShadow = `inset 0 0 0 ${
     themeData["checkbox.borderWidth"]
   } 
-    ${themeData["checkbox.borderColor"]}
+    ${borderColor}
   `;
   let outerShadow = "0 0 0 0px transparent";
 
@@ -93,7 +97,7 @@ export default function stylesheet(props, themeData, theme) {
     // do nothing
   } else if (isPressed) {
     outerShadow = `0 0 0 ${themeData["checkbox.pressed.halo.width"]} ${
-      themeData["checkbox.focus.halo.color"]
+      themeData["checkbox.pressed.halo.color"]
     }`;
   } else if (hasHover) {
     outerShadow = `0 0 0 ${themeData["checkbox.hover.halo.width"]} ${