+++ packages/theme-data/src/colorSchemes/darkGray/unresolvedRoles.js
@@ -2,6 +2,7 @@ import extendTheme from "../../utils/extendTheme";
 import mapKeys from "../../utils/mapKeys";
 import darkBlueTheme from "../darkBlue";
 import system from "./system";
+import button from "./components/button";
 import divider from "./components/divider";
 import iconButton from "./components/iconButton";
 import input from "./components/input";
@@ -13,6 +14,7 @@ const darkGrayThemeConfig = extendTheme(
   Object.assign(
     {},
     mapKeys(system.colorScheme, key => `colorScheme.${key}`),
+    button,
     divider,
     iconButton,
     input,