+++ packages/theme-data/CHANGELOG.md
@@ -1,3 +1,34 @@
+# [@hig/theme-data-v2.8.0](https://github.com/Autodesk/hig/compare/@hig/theme-data@2.7.0...@hig/theme-data@2.8.0) (2019-06-07)
+
+
+### Bug Fixes
+
+* correct base color for dark blue theme ([7bea1bf](https://github.com/Autodesk/hig/commit/7bea1bf))
+* correct button hover/focus colors ([302e984](https://github.com/Autodesk/hig/commit/302e984))
+* correct checkbox hover/focus colors ([7c56ed7](https://github.com/Autodesk/hig/commit/7c56ed7))
+* correct slider hover/focus colors ([fab82c9](https://github.com/Autodesk/hig/commit/fab82c9))
+
+
+### Features
+
+* export theme-data as less variables ([307112d](https://github.com/Autodesk/hig/commit/307112d))
+
+# [@hig/theme-data-v2.7.0](https://github.com/Autodesk/hig/compare/@hig/theme-data@2.6.0...@hig/theme-data@2.7.0) (2019-05-22)
+
+
+### Bug Fixes
+
+* use correct input height ([02af709](https://github.com/Autodesk/hig/commit/02af709))
+
+
+### Features
+
+* new disabled opacity ([2e9e88d](https://github.com/Autodesk/hig/commit/2e9e88d))
+* new focus state border color for checkbox and radio-button ([07c6310](https://github.com/Autodesk/hig/commit/07c6310))
+* new opacity for focus state halos ([a4830ef](https://github.com/Autodesk/hig/commit/a4830ef))
+* new pressed state halo colors ([dc6e337](https://github.com/Autodesk/hig/commit/dc6e337))
+* new text-link focus halo color ([97d8761](https://github.com/Autodesk/hig/commit/97d8761))
+
 # [@hig/theme-data-v2.6.0](https://github.com/Autodesk/hig/compare/@hig/theme-data@2.5.0...@hig/theme-data@2.6.0) (2019-05-17)